Add HideProfileBubble.kt 剥离隐藏点赞提醒
